# Conflicts:
#	yuepei-admin/src/main/java/com/yuepei/web/agent/service/impl/AgentServiceImpl.java
#	yuepei-applet/src/main/java/com/yuepei/controller/wechat/user/LoginController.java
#	yuepei-system/src/main/java/com/yuepei/system/domain/Device.java
#	yuepei-system/src/main/java/com/yuepei/system/mapper/DeviceMapper.java
#	yuepei-system/src/main/java/com/yuepei/system/mapper/SysUserMapper.java
#	yuepei-system/src/main/java/com/yuepei/system/service/impl/HospitalDeviceServiceImpl.java
#	yuepei-system/src/main/resources/mapper/device/DeviceMapper.xml
#	yuepei-system/src/main/resources/mapper/system/SysUserMapper.xml

+++ yuepei-system/src/main/java/com/yuepei/system/domain/UserLeaseOrder.java
@@ -72,7 +72,7 @@ public class UserLeaseOrder extends BaseEntity {
     private String payType;
 
     /** 实付金额 */
-    private Long netAmount;
+    private BigDecimal netAmount;
 
     /** 优惠券额度 */
     private Long couponPrice;
